requirements.txt
streamlit==1.36.0
google-generativeai
python-dotenv
SpeechRecognition
pyttsx3